gaps detection extraction from main
